app/indexers/requirements.txt
aiocache==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hexbytes==0.3.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
jsonschema==4.17.3
multidict==6.0.4
parsimonious==0.9.0
polars==0.17.2
psycopg2==2.9.6
pycryptodome==3.17
pydantic==1.10.7
pyrsistent==0.19.3
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a33
sqlmodel==0.0.8
toolz==0.12.0
typing-extensions==4.5.0
urllib3==1.26.15
web3==6.2.0
websockets==11.0.1
yarl==1.8.2
